+++ saphyr/src/loader.rs
@@ -372,6 +372,9 @@ pub(crate) fn parse_f64(v: &str) -> Option<f64> {
         ".inf" | ".Inf" | ".INF" | "+.inf" | "+.Inf" | "+.INF" => Some(f64::INFINITY),
         "-.inf" | "-.Inf" | "-.INF" => Some(f64::NEG_INFINITY),
         ".nan" | ".NaN" | ".NAN" => Some(f64::NAN),
-        _ => v.parse::<f64>().ok(),
+        // Test that `v` contains a digit so as not to pass in strings like `inf`,
+        // which rust will parse as a float.
+        _ if v.as_bytes().iter().any(u8::is_ascii_digit) => v.parse::<f64>().ok(),
+        _ => None,
     }
 }