+++ saphyr/tests/basic.rs
@@ -248,3 +248,42 @@ fn test_integer_key() {
     let first = out.into_iter().next().unwrap();
     assert_eq!(first[0]["important"].as_bool().unwrap(), true);
 }
+
+#[test]
+fn test_nominal_float_parse() {
+    use std::fmt::Write;
+
+    // back test for
+    // https://github.com/Ethiraric/yaml-rust2/issues/50
+
+    // Test cases derived from https://doc.rust-lang.org/std/primitive.f64.html#impl-FromStr-for-f64
+
+    // Generates a document like so:
+    // ```yaml
+    // - +nan
+    // - -nan
+    // - nan
+    // - +NAN
+    // ...
+    // ```
+    // Every single one of these values should be parsed as a string in yaml,
+    // but would be parsed as a float according to rust. This test verifies they
+    // all end up parsed as strings.
+    let raw = &["nan", "NAN", "NaN", "inf", "infinity", "Infinity"]
+        .iter()
+        .fold(String::new(), |mut output, base| {
+            write!(output, "- +{base}\n- -{base}\n- {base}\n").unwrap();
+            output
+        });
+
+    println!("parsing {raw}");
+
+    let doc = Yaml::load_from_str(raw).expect("could not parse document");
+    let doc = doc.into_iter().next().unwrap();
+
+    assert!(doc.is_sequence());
+
+    for it in doc {
+        assert!(it.as_str().is_some());
+    }
+}